+++ src/test/DijkstraTest.java
@@ -32,7 +32,7 @@ public class DijkstraTest {
         graph.calculateShortestPaths(4);
         List<Integer> path = algorithm.getPath(4, 2);
         assertNull(path);
-        assertEquals(Integer.MAX_VALUE, algorithm.getCost(4, 2));
+        assertEquals(Long.MAX_VALUE, algorithm.getCost(4, 2));
     }
 
     @Test
@@ -102,7 +102,7 @@ public class DijkstraTest {
         graph.calculateShortestPaths(95);
         List<Integer> path = algorithm.getPath(95, 0);
         assertNull(path);
-        assertEquals(Integer.MAX_VALUE, algorithm.getCost(95, 0));
+        assertEquals(Long.MAX_VALUE, algorithm.getCost(95, 0));
     }
 
     @Test
@@ -125,6 +125,6 @@ public class DijkstraTest {
         graph.calculateShortestPaths(2);
         List<Integer> path = algorithm.getPath(1, 0);
         assertNull(path);
-        assertEquals(Integer.MAX_VALUE, algorithm.getCost(1, 0));
+        assertEquals(Long.MAX_VALUE, algorithm.getCost(1, 0));
     }
 }
